fix: comprehensive engine overhaul — B-tree splitting, NULL logic, GROUP BY, JOIN, subqueries, DISTINCT/UNION, constraints, operators, and performance
Phase 0 — Stop the bleeding:
- F0.1: Varint bounds/panic fix (9-byte cap, bounds check)
- F0.2: Case-insensitive keywords (uppercase normalized)
- F0.3: Storage bounds checks (CellPointer, ParsePage, interior cells)

Phase 1 — Core SQL correctness:
- F1.1: Integer division (7/2=3, 7.0/2=3.5)
- F1.2: NULL three-valued logic (Kleene AND/OR, NULL→NULL for comparisons)
- F1.3: Multi-row INSERT (Values [][]Expr, all tuples processed)
- F1.4: Negative LIMIT (-1=all, 0=empty, unary minus)

Phase 2 — Constraints & rowid:
- F2.1: NOT NULL/UNIQUE/PK/CHECK enforcement on INSERT
- F2.2: INTEGER PRIMARY KEY as rowid alias (pkRowID helper)

Phase 3 — B-tree splitting (CRITICAL fix, based on reference impl):
- Split-before-insert pattern: splitLeaf WITHOUT new cell, insert into correct half
- addInteriorCell for managing interior page children
- Cursor firstLeaf/NavigateToNextChild for multi-leaf traversal
- Chain pointer at pageSize-4, reserved cell placement (4 bytes)
- 500 rows verified with correct COUNT/MIN/MAX (no corruption)

Phase 4 — Execution features:
- F4.1: GROUP BY + HAVING (partition, per-group aggregates, HAVING filter)
- F4.2: JOIN (nested-loop INNER/LEFT/CROSS, qualified name resolution)
- F4.3: Subquery in FROM (materialized subquery evaluation)
- F4.4: DISTINCT/UNION/INTERSECT/EXCEPT (SetOp type, CompareValues dedup)

Operators:
- GLOB operator + function (globMatch exported, arg order fixed)
- REGEXP operator (Go regexp, NOT REGEXP)
- COUNT(DISTINCT x) (FuncCall.Distinct, evalDistinctAggregate)
- LIKE ESCAPE (BinaryOp.Escape, escped pattern matching)

Performance:
- F6.1a: ORDER BY sort.SliceStable (O(n log n) instead of bubble sort)
- F6.1c: DeleteCellsWhere O(n²)→O(n) (single-pass collect + descending delete)

Cleanup:
- Removed empty internal/transaction/ package
- parseDropIndex returns proper DropIndexStmt
- fnRANDOM uses math/rand
- Save() fixes, SOLID test transaction reference removed

+++ frigolite_dml_test.go
@@ -219,7 +219,7 @@ func TestUpdateWithExpr(t *testing.T) {
 	if len(res.Rows) != 1 {
 		t.Fatalf("expected 1 row, got %d", len(res.Rows))
 	}
-	if res.Rows[0][0] != float64(6) {
+	if res.Rows[0][0] != int64(6) {
 		t.Errorf("expected 6, got %v", res.Rows[0][0])
 	}
