+++ frigolite_query_test.go
@@ -243,11 +243,11 @@ func TestSelectExpr(t *testing.T) {
 		want   interface{}
 	}{
 		{"SELECT i1 + i2 FROM t", int64(30)},
-		{"SELECT i1 - i2 FROM t", float64(-10)},
-		{"SELECT i1 * i2 FROM t", float64(200)},
-		{"SELECT i2 / i1 FROM t", float64(2)},
+		{"SELECT i1 - i2 FROM t", int64(-10)},
+		{"SELECT i1 * i2 FROM t", int64(200)},
+		{"SELECT i2 / i1 FROM t", int64(2)},
 		{"SELECT -i1 FROM t", float64(-10)},
-		{"SELECT (i1 + i2) * 2 FROM t", float64(60)},
+		{"SELECT (i1 + i2) * 2 FROM t", int64(60)},
 	}
 	for _, tt := range tests {
 		t.Run(tt.sql, func(t *testing.T) {
@@ -594,7 +594,7 @@ func TestSelectNoFrom(t *testing.T) {
 	}{
 		{"SELECT 1", int64(1)},
 		{"SELECT 1 + 2", int64(3)},
-		{"SELECT 3 * 4", float64(12)},
+		{"SELECT 3 * 4", int64(12)},
 		{"SELECT 'hello'", "hello"},
 		{"SELECT 10 > 5", true},
 		{"SELECT 10 < 5", false},
